+++ lib/bloc.dart
@@ -1,4 +1,8 @@
+import 'dart:io';
+
+import 'package:device_info/device_info.dart';
 import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
 import 'package:lighthouse_pm/data/Database.dart';
 import 'package:lighthouse_pm/data/bloc/ViveBaseStationBloc.dart';
 import 'package:lighthouse_pm/lighthouseProvider/LighthousePowerState.dart';
@@ -48,11 +52,16 @@ class SettingsBloc {
 
   SettingsBloc(this.db);
 
+  // region IDS
+  static const SCAN_DURATION_VALUES = const [5, 10, 15, 20];
+
   //IDS
   static const DEFAULT_SLEEP_STATE_ID = 1;
   static const VIVE_BASE_STATION_ENABLED_ID = 2;
   static const SCAN_DURATION_ID = 3;
-  static const SCAN_DURATION_VALUES = const [5, 10, 15, 20];
+  static const PREFERED_THEME_ID = 4;
+
+  // endregion
 
   Stream<LighthousePowerState> getSleepStateAsStream(
       {LighthousePowerState defaultValue = LighthousePowerState.SLEEP}) {
@@ -124,5 +133,58 @@ class SettingsBloc {
     return db.into(db.simpleSettings).insert(
         SimpleSetting(id: SCAN_DURATION_ID, data: duration.toRadixString(10)),
         mode: InsertMode.insertOrReplace);
+  }
+
+  Stream<ThemeMode> getPreferedThemeAsStream() {
+    return (db.select(db.simpleSettings)
+          ..where((tbl) => tbl.id.equals(PREFERED_THEME_ID)))
+        .watch()
+        .asyncMap((event) async {
+      if (event.length == 1 && event[0].data != null) {
+        final themeIndex = int.tryParse(event[0].data, radix: 10);
+        if (themeIndex >= 0 && themeIndex < ThemeMode.values.length) {
+          var themeMode = ThemeMode.values[themeIndex];
+          if (themeMode == ThemeMode.system && !await supportsThemeModeSystem) {
+            themeMode = ThemeMode.light;
+          }
+          return themeMode;
+        }
+      }
+      return await defaultThemeMode;
+    });
+  }
+
+  Future<void> setPreferedTheme(ThemeMode theme) {
+    return db.into(db.simpleSettings).insert(
+        SimpleSetting(
+            id: PREFERED_THEME_ID, data: '${theme.index.toRadixString(10)}'),
+        mode: InsertMode.insertOrReplace);
+  }
+
+  /// Check if the currently running device supports system mode theme.
+  static Future<bool> get supportsThemeModeSystem async {
+    final DeviceInfoPlugin deviceInfo = DeviceInfoPlugin();
+    if (Platform.isAndroid) {
+      final android = await deviceInfo.androidInfo;
+      if (android.version.sdkInt >= 29 /* Android 10 */) {
+        return true;
+      }
+    } else if (Platform.isIOS) {
+      final ios = await deviceInfo.iosInfo;
+      // TODO: check what systemVersion returns on an actual iOS device.
+      if (int.tryParse(ios.systemVersion, radix: 10) >= 13 /* iOS 13*/) {
+        return true;
+      }
+    }
+    return false;
+  }
+
+  /// Get the default theme mode taking [supportsThemeModeSystem]
+  /// into consideration.
+  static Future<ThemeMode> get defaultThemeMode async {
+    if (await supportsThemeModeSystem) {
+      return ThemeMode.system;
+    }
+    return ThemeMode.light;
   }
 }